Reformatting repo to front/back end folders
