+++ MusiQueue-android/app/src/main/res/layout/activity_queue.xml
