+++ MusiQueue-android/app/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="app_name">MusiQueue</string>
+    <string name="title_activity_queue">Queue</string>
+</resources>
