--- app/src/main/res/layout/activity_main.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.vince.youtubeplayertest.Activities.MainActivity">
-
-    <Button
-        android:id="@+id/create_queue_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="CREATE QUEUE"
-        android:layout_centerInParent="true" />
-
-</RelativeLayout>
